feat(create-profile):  add a protected create-profile route and refactor ProfileForm
