+++ app/create-profile/page.jsx
@@ -1,0 +1,10 @@
+"use client";
+
+import ProfileForm from "@components/ProfileForm/ProfileForm";
+
+function CreateProfile() {
+
+  return <ProfileForm />
+}
+
+export default CreateProfile
