+++ app/profile/profile.module.css
@@ -5,6 +5,48 @@
   gap: 2rem;
 }
 
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  height: 100%;
+}
+
+.buttons_wrapper{
+  display: flex;
+  width: 100%;
+  gap: 10px;
+}
+
+.button {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  color: var(--primary-light-color);
+  font-size: var(--font-xl);
+  font-weight: var(--font-black);
+}
+
+.edit_button {
+  background-color: transparent;
+  border: 2px solid var(--primary-light-color);
+}
+
+.validate_button {
+  background-color: var(--primary-accent-color);
+}
+
+.display_creds {
+  background-color: transparent;
+  margin-right: auto;
+  color: var(--primary-light-color);
+  padding: 0;
+}
+
+.visible_button {
+  background-color: transparent;
+}
+
 .button {
   display: flex;
   flex-direction: row;
